src/requirements.txt
django==4.2.10
celery==5.3.6
flower==2.0.1
pillow==10.2.0
djangorestframework==3.14.0
psycopg==3.1.18
psycopg2
pytest-django==4.8.0
django-redis
Faker
django-cors-headers
setuptools
wheel
gunicorn
uvicorn
redis
virtualenv
django-debug-toolbar
debugpy
django_extensions
pandas
numpy